Split TournamentManager and add GameStats
